+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] / 2023-12-07
+### Updated
+- Add `DesignAutomationEngineDateUtils` to enable `DeprecationDate` on Engine.
+- Update `ForgeCustomHeaderValueHandler` with content and only when engine is deprecated.
+- Update `UpdateCustomHeaderEngineName` to check if engine is deprecated.
+
 ## [1.0.7] / 2023-10-03
 ### Updated
 - Update `ParameterActivityLanguageAttribute` to ignore empty value (Fix: #9)
@@ -42,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release (Copy project from: [RevitAddin.DA.Tester](https://github.com/ricaun-io/RevitAddin.DA.Tester/tree/package))
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.8]: ../../compare/1.0.7...1.0.8
 [1.0.7]: ../../compare/1.0.6...1.0.7
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5